Derive Macro derive_discriminant::ToDiscriminant

#[derive(ToDiscriminant)]
{
    // Attributes available to this derive:
    #[sub_discriminant]
    #[discriminant_attr]
}

Derive the ToDiscriminant trait and create a <Type Name>Discriminant enum

This derive macro is enum-only.

The discriminant enum is a copy of the input enum with all fields of every variant removed.
*) The exception to that rule is the #[child] attribute

Helper attributes

#[sub_discriminant]: only usable on variants with a single field; instead of no fields, the discriminant of the single field will be included in the discriminant, acting as a sub-discriminant.

#[discriminant_attr(…)]: usable on the enum itself or on any variant; applies #[…] in its place on the discriminant.

Attributes on the Discriminant

All attributes on variants and the type itself are cleared when constructing the discriminant. If the discriminant is supposed to also have an attribute, you must double it with #[discriminant_attr(…)]

Example

#[derive(ToDiscriminant)]
#[discriminant_attr(derive(Debug, Eq, PartialEq))]
pub enum EnumA {
    A(u8),
    #[sub_discriminant]
    B(EnumB)
}

#[derive(ToDiscriminant)]
#[discriminant_attr(derive(Debug, Eq, PartialEq))]
#[discriminant_attr(repr(u8))]
pub enum EnumB {
    Foo(u8),
    Bar(String),
    #[cfg(feature = "some-feature")]
    #[discriminant_attr(cfg(feature = "some-feature"))]
    WindowsBar(OsString)
}

let a = EnumA::A(1);
assert_eq!(a.to_discriminant(), EnumADiscriminant::A);
let b = EnumA::B(EnumB::Bar("bar".to_string()));
assert_eq!(b.to_discriminant(), EnumADiscriminant::B(EnumBDiscriminant::Bar));